pyproject.toml
[tool.poetry]
name = "mini-erp-gestao-vendas"
version = "0.1.0"
description = "gestao de vendas"
authors = ["jose felipe <felipe.brx.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "mini_erp_gestao_vendas"}]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.5.2"
boto3 = "1.24.89"
botocore = "1.27.89"
certifi = "2021.10.8"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
cryptography = "36.0.1"
defusedxml = "0.7.1"
dj-database-url = "0.4.2"
dj-static = "0.0.6"
Django = "3.2.9"
django-bootstrap-form = "3.4"
django-debug-toolbar = "3.7.0"
django-model-utils = "4.2.0"
django-storages = "1.13.1"
docutils = "0.19"
idna = "3.4"
jmespath = "1.0.1"
oauthlib = "3.2.1"
Pillow = "^9.2.0"
psycopg2 = "2.9.4"
pycparser = "2.21"
PyJWT = "2.5.0"
python-dateutil = "2.8.2"
python-decouple = "3.6"
python3-openid = "3.2.0"
pytz = "2022.4"
s3transfer = "0.6.0"
static3 = "0.7.0"
django-jazzmin = "2.5.0"
django-filter = "22.1"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
six = "1.16.0"
sqlparse = "0.4.3"
urllib3 = "1.26.12"
uWSGI = "^2.0.20"
django-cors-headers = "^3.13.0"
django-hosts = "^5.2"
django-role-permissions = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
